requirements.txt
dspy>=2.6.10
pandas>=2.2.3
tqdm>=4.67.1
python-dotenv==1.0.1
rouge>=1.0.0
